+++ CHANGELOG.md
@@ -1,3 +1,34 @@
+## [1.11.4](https://github.com/knowledgepixels/nanopub-registry/compare/nanopub-registry-1.11.3...nanopub-registry-1.11.4) (2026-08-05)
+
+### Bug Fixes
+
+* let the trust calculation ignore the setting's invalidation, not the record ([#60](https://github.com/knowledgepixels/nanopub-registry/issues/60)) ([2a004ed](https://github.com/knowledgepixels/nanopub-registry/commit/2a004eda2198c24ffa49640a7b7dba6335a760dd)), closes [#125](https://github.com/knowledgepixels/nanopub-registry/issues/125) [#125](https://github.com/knowledgepixels/nanopub-registry/issues/125) [#125](https://github.com/knowledgepixels/nanopub-registry/issues/125) [#125](https://github.com/knowledgepixels/nanopub-registry/issues/125)
+
+### General maintenance
+
+* setting next snapshot version [skip ci] ([b67b1c7](https://github.com/knowledgepixels/nanopub-registry/commit/b67b1c71c1b6885af83a571e7b240af562776e04))
+
+## [1.11.3](https://github.com/knowledgepixels/nanopub-registry/compare/nanopub-registry-1.11.2...nanopub-registry-1.11.3) (2026-08-05)
+
+### Bug Fixes
+
+* make the bootstrap setting immune to invalidation ([#60](https://github.com/knowledgepixels/nanopub-registry/issues/60)) ([abe932e](https://github.com/knowledgepixels/nanopub-registry/commit/abe932ec63040cda5e4a9ee7d2633cd0ab90f93e))
+* narrow [#60](https://github.com/knowledgepixels/nanopub-registry/issues/60) immunity to trust edges only ([d530efa](https://github.com/knowledgepixels/nanopub-registry/commit/d530efaca698c9da1d2058bdf7864d7d5bb9509d))
+
+### General maintenance
+
+* setting next snapshot version [skip ci] ([fe3e93c](https://github.com/knowledgepixels/nanopub-registry/commit/fe3e93c66fcb2bf371fd26647edda1b7d980dc4c))
+
+## [1.11.2](https://github.com/knowledgepixels/nanopub-registry/compare/nanopub-registry-1.11.1...nanopub-registry-1.11.2) (2026-08-05)
+
+### Dependency updates
+
+* **core-deps:** update org.nanopub:nanopub dependency to v1.91.0 ([e5ed6d4](https://github.com/knowledgepixels/nanopub-registry/commit/e5ed6d4623dbefdc0be06fa9f74401e26c7d120f))
+
+### General maintenance
+
+* setting next snapshot version [skip ci] ([e86e484](https://github.com/knowledgepixels/nanopub-registry/commit/e86e484a02f143899456be12024259285d361236))
+
 ## [1.11.1](https://github.com/knowledgepixels/nanopub-registry/compare/nanopub-registry-1.11.0...nanopub-registry-1.11.1) (2026-06-30)
 
 ### Bug Fixes